Conform to rules

+++ bootloader/inc/frame.h
@@ -8,7 +8,7 @@
 // Types for frames that will be the *unencrypted* version of incoming data.
 // These should not be indexed; utilize their members.
 typedef struct begin_frame_t {
-    uint8_t version;
+    uint16_t version;
     uint16_t num_packets;
     uint16_t bytesize;
     uint8_t signature[HMAC_SIG_LENGTH];
